Move navbar links into separate include
This means that if someone wants full control over the links that appear
in their navbar, they can just override the _includes/navbar-links.html
file in their site.

I considered allowing for a 'navbar' data structure to be specified in
the site's configuration, but I think this is a more complicated
approach, and decided to ignore it in favour of allowing the links to be
overriden.